+++ api/package.json
@@ -30,7 +30,6 @@
     "egg-redis": "^2.4.0",
     "egg-scripts": "^2.6.0",
     "egg-session": "^3.3.0",
-    "egg-socket.io": "^4.1.6",
     "immer": "^9.0.12",
     "lodash": "^4.17.21",
     "ot-json1": "^1.0.2",
