+++ api/typings/app/controller/index.d.ts
@@ -3,14 +3,12 @@
 
 import 'egg';
 import ExportCollDoc from '../../../app/controller/collDoc';
-import ExportDoc from '../../../app/controller/doc';
 import ExportMdDoc from '../../../app/controller/mdDoc';
 import ExportUser from '../../../app/controller/user';
 
 declare module 'egg' {
   interface IController {
     collDoc: ExportCollDoc;
-    doc: ExportDoc;
     mdDoc: ExportMdDoc;
     user: ExportUser;
   }
