+++ web/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@arco-design/web-react": "^2.32.0",
+    "ahooks": "^3.7.1",
     "axios": "^0.21.4",
     "bson-objectid": "^2.0.3",
     "classnames": "^2.3.1",
