+++ web/src/pages/coll_doc/components/Elements/Link/index.less
@@ -1,0 +1,20 @@
+.link-op-container {
+  padding: 4px 6px;
+  border: 1px solid rgb(var(--gray-4));
+  border-radius: 4px;
+  background-color: #fff;
+  display: flex;
+  gap: 2px;
+
+  .op {
+    display: flex;
+    font-size: 18px;
+    cursor: pointer;
+    padding: 2px;
+    border-radius: 2px;
+
+    &:hover {
+      background-color: rgb(var(--gray-3));
+    }
+  }
+}
