+++ web/src/pages/coll_doc/components/Elements/index.tsx
@@ -1,0 +1,6 @@
+import DefaultElement from './DefaultElement';
+import Leaf from './Leaf';
+import Link from './Link';
+import Paragraph from './Paragraph';
+
+export { DefaultElement, Leaf, Link, Paragraph };
