+++ web/src/pages/coll_doc/components/HoveringToolbar/ColorSelect/index.less
@@ -1,0 +1,73 @@
+.color-select {
+  display: flex;
+  align-items: center;
+  border-radius: 4px;
+  padding: 0 2px;
+
+  .color-set {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    cursor: pointer;
+    border-radius: 4px;
+
+    .icon {
+      font-size: 20px;
+      display: flex;
+    }
+
+    .color-preview {
+      height: 2px;
+      width: 12px;
+      margin: 0 auto 2px;
+      border-radius: 4px;
+      background-color: #000;
+    }
+
+    &:hover {
+      background-color: rgb(var(--gray-3));
+    }
+  }
+
+  .caret-down {
+    cursor: pointer;
+    border-radius: 4px;
+
+    &:hover {
+      background-color: rgb(var(--gray-3));
+    }
+
+    .color-container {
+      padding: 10px 10px 0px;
+      background: rgb(255, 255, 255);
+      border-radius: 4px;
+      border: 1px solid rgb(var(--gray-3));
+      position: absolute;
+
+      .clear-color {
+        padding: 6px 4px;
+        border-radius: 4px;
+        display: flex;
+        align-items: center;
+        margin-bottom: 4px;
+        gap: 6px;
+
+        &:hover {
+          background-color: rgb(var(--gray-2));
+        }
+      }
+
+      .sketch-picker {
+        padding: 0 !important;
+        border: none;
+        box-shadow: none !important;
+        width: 240px !important;
+        pointer-events: all !important;
+      }
+    }
+  }
+}
+
+.active {
+  background-color: rgb(var(--gray-3));
+}
