+++ web/src/pages/coll_doc/components/HoveringToolbar/FontSize/index.less
