+++ web/src/pages/coll_doc/components/HoveringToolbar/index.less
