+++ web/src/pages/coll_doc/components/IconFont/index.tsx
@@ -1,0 +1,7 @@
+import { Icon } from '@arco-design/web-react';
+
+const IconFont = Icon.addFromIconFontCn({
+  src: '//at.alicdn.com/t/c/font_3601028_fn7l7occxh.js',
+});
+
+export default IconFont;
