+++ web/src/pages/coll_doc/plugins/withWebSocket.ts
