--- web/src/pages/edit/index.less
@@ -1,9 +1,0 @@
-.edit-wrapper {
-  padding: 20px 0;
-  display: flex;
-  justify-content: center;
-
-  .edit-container {
-    width: 800px;
-  }
-}
