+++ web/src/pages/home/index.tsx
@@ -1,5 +1,3 @@
-import { createDoc, searchDoc } from '@/services';
-import { TDoc } from '@/typings/doc';
 import {
   Button,
   Avatar,
@@ -16,11 +14,11 @@ import {
 import { useEffect, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import dayjs from 'dayjs';
-import './index.less';
 import { createMdDoc, searchMdDoc } from '@/services/mdDoc';
 import { EDocType, TMdDoc } from '@/typings/mdDoc';
 import { createCollDoc, getCollDocList } from '@/services/collDoc';
 import { CollDoc } from '@/typings/collDoc';
+import './index.less';
 
 const { Item: ListItem } = List;
 const { Group: RadioGroup } = Radio;
@@ -28,7 +26,7 @@ const { Item: FormItem, useForm } = Form;
 
 export default function Home() {
   const history = useHistory();
-  const [docs, setDocs] = useState<(TDoc | TMdDoc | CollDoc)[]>();
+  const [docs, setDocs] = useState<(TMdDoc | CollDoc)[]>();
   const [loading, setLoading] = useState(true);
   const [visible, setVisible] = useState(false);
   const [docType, setDocType] = useState<EDocType>(EDocType.CollDocs);
@@ -38,9 +36,6 @@ export default function Home() {
   useEffect(() => {
     let searchService;
     switch (docType) {
-      case EDocType.Docs:
-        searchService = searchDoc;
-        break;
       case EDocType.MarkDown:
         searchService = searchMdDoc;
         break;
@@ -57,30 +52,6 @@ export default function Home() {
       });
   }, [docType]);
 
-  function handleCreateDoc() {
-    createDoc({})
-      .then((res) => {
-        const { _id: docId } = res;
-        history.push(`/edit/${docId}`);
-      })
-      .catch((e: Error) => {
-        Message.error(e.message || '创建失败');
-        console.error(e);
-      });
-  }
-
-  function handleCreateCollDoc() {
-    createCollDoc({})
-      .then((res) => {
-        const { _id: docId } = res;
-        history.push(`/coll_doc/${docId}`);
-      })
-      .catch((e: Error) => {
-        Message.error(e.message || '创建失败');
-        console.error(e);
-      });
-  }
-
   function openModal() {
     setVisible(true);
   }
@@ -96,12 +67,19 @@ export default function Home() {
     }
   }
 
+  async function handleCreateCollDoc() {
+    try {
+      const { _id: docId } = await createCollDoc({});
+      history.push(`/coll_doc/${docId}`);
+    } catch (error) {
+      Message.error('创建失败');
+      console.error(error);
+    }
+  }
+
   function handleEditDoc(docId: string) {
     let url = '';
     switch (docType) {
-      case EDocType.Docs:
-        url = `/edit/${docId}`;
-        break;
       case EDocType.MarkDown:
         url = `/md_edit/${docId}`;
         break;
@@ -115,18 +93,18 @@ export default function Home() {
   }
 
   return (
-    <div className="home-container">
-      <div className="header-container">
+    <div className='home-container'>
+      <div className='header-container'>
         <Avatar size={36}>
-          <img src={window.userInfo?.avatar} alt="avatar" />
+          <img src={window.userInfo?.avatar} alt='avatar' />
         </Avatar>
       </div>
-      <div className="main-title-and-create">
-        <span className="main-title">主页</span>
+      <div className='main-title-and-create'>
+        <span className='main-title'>主页</span>
 
         <Dropdown
-          position="br"
-          trigger="click"
+          position='br'
+          trigger='click'
           droplist={
             <Menu>
               <Menu.Item key={EDocType.CollDocs} onClick={handleCreateCollDoc}>
@@ -135,33 +113,29 @@ export default function Home() {
               <Menu.Item key={EDocType.MarkDown} onClick={openModal}>
                 MarkDown
               </Menu.Item>
-              <Menu.Item key={EDocType.Docs} onClick={handleCreateDoc} disabled>
-                文档 2.0
-              </Menu.Item>
             </Menu>
-          }
-        >
-          <Button type="primary">新建</Button>
+          }>
+          <Button type='primary'>新建</Button>
         </Dropdown>
       </div>
-      <div className="banner">
-        <div className="left">
-          <p className="title">好记性不如烂笔头，写个文档记录一下吧!</p>
-          <p className="desc">
+      <div className='banner'>
+        <div className='left'>
+          <p className='title'>好记性不如烂笔头，写个文档记录一下吧!</p>
+          <p className='desc'>
             人生要用简单的心境，对待复杂的人生，最无情的不是人，是时间；最珍贵的不是金钱，是情感；最可怕的不是失恋，是心身不全；最舒适的不是酒店，是家里；最难听的不是脏话，是无言；最美好的不是未来，是今天。
           </p>
         </div>
-        <div className="right">
+        <div className='right'>
           <img
-            className="banner-img"
+            className='banner-img'
             src={require('@/assets/banner.png')}
-            alt=""
+            alt=''
           />
         </div>
       </div>
-      <div className="recent-list">
-        <div className="recent-header">
-          <span className="title">最近访问</span>
+      <div className='recent-list'>
+        <div className='recent-header'>
+          <span className='title'>最近访问</span>
           <RadioGroup
             value={docType}
             onChange={setDocType}
@@ -174,14 +148,9 @@ export default function Home() {
                 label: '协同文档',
                 value: EDocType.CollDocs,
               },
-              {
-                label: '文档 2.0',
-                value: EDocType.Docs,
-              },
-            ]}
-          ></RadioGroup>
+            ]}></RadioGroup>
         </div>
-        <div className="list-title row">
+        <div className='list-title row'>
           <span>文档名称</span>
           <span>文档所有人</span>
           <span>更新时间</span>
@@ -189,14 +158,13 @@ export default function Home() {
         <Spin loading={loading} block>
           <List
             dataSource={docs}
-            wrapperClassName="list"
+            wrapperClassName='list'
             render={(item, index) => (
               <ListItem
                 key={index}
-                className="list-item"
-                onClick={() => handleEditDoc(item._id)}
-              >
-                <div className="row">
+                className='list-item'
+                onClick={() => handleEditDoc(item._id)}>
+                <div className='row'>
                   <span>
                     <img
                       src={require('@/assets/doc.svg')}
@@ -210,24 +178,21 @@ export default function Home() {
                   </span>
                 </div>
               </ListItem>
-            )}
-          ></List>
+            )}></List>
         </Spin>
       </div>
       <Modal
-        title="文档信息"
+        title='文档信息'
         visible={visible}
         onCancel={() => setVisible(false)}
         onOk={handleCreateMdDoc}
-        unmountOnExit
-      >
+        unmountOnExit>
         <Form form={form}>
           <FormItem
-            label="文档标题"
-            field="title"
-            rules={[{ required: true, message: '请输入文档标题' }]}
-          >
-            <Input placeholder="请输入" />
+            label='文档标题'
+            field='title'
+            rules={[{ required: true, message: '请输入文档标题' }]}>
+            <Input placeholder='请输入' />
           </FormItem>
         </Form>
       </Modal>
