--- web/src/pages/tmp_doc/components/IconFont/index.tsx
@@ -1,7 +1,0 @@
-import { Icon } from '@arco-design/web-react';
-
-const IconFont = Icon.addFromIconFontCn({
-  src: '//at.alicdn.com/t/c/font_3601028_6bghmsbh3cg.js',
-});
-
-export default IconFont;
