--- web/src/pages/tmp_doc/index.less
@@ -1,18 +1,0 @@
-#coll-doc-container {
-  height: 100%;
-  width: 1000px;
-  margin: 0 auto;
-  overflow: hidden;
-
-  .editor {
-    padding: 20px 0;
-  }
-}
-
-.loading-bar {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  justify-content: center;
-  align-items: center;
-}
