+++ web/src/typings/collDoc.ts
@@ -1,8 +1,7 @@
-
 /**
  * 文档结构
  */
- export interface CollDoc {
+export interface CollDoc {
   /** 文档 id */
   _id: string;
   /** 文档名称 */
@@ -13,4 +12,4 @@
   createdAt: number;
   /** 更新时间 unix 时间戳 */
   updatedAt: number;
-};
+}
